# Noordle/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Noordle, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/api/Game": {
      "post": {
        "tags": [
          "Game"
        ],
        "operationId": "Game_POST",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartGameDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/StartGameDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/StartGameDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/StartGameResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartGameResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartGameResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/Game/{gameId}/{guess}": {
      "put": {
        "tags": [
          "Game"
        ],
        "operationId": "Game_PUT",
        "parameters": [
          {
            "name": "gameId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "guess",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GuessResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuessResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuessResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/Game/{gameId}": {
      "delete": {
        "tags": [
          "Game"
        ],
        "operationId": "Game_DELETE",
        "parameters": [
          {
            "name": "gameId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "GuessResponse": {
        "type": "object",
        "properties": {
          "isValid": {
            "type": "boolean"
          },
          "matches": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WordMatch"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "LetterStatus": {
        "enum": [
          "Correct",
          "IncorrectLocation",
          "Incorrect"
        ],
        "type": "string"
      },
      "StartGameDto": {
        "type": "object",
        "properties": {
          "boardCount": {
            "type": "integer",
            "format": "int32"
          },
          "wordLength": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "StartGameResponse": {
        "type": "object",
        "properties": {
          "gameId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
      "WordMatch": {
        "type": "object",
        "properties": {
          "letters": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LetterStatus"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}